Move FromMessageError to main error module.
